requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
pytesseract
pillow
pandas
openpyxl
pyyaml
